BIL-3801 УУ. Оптимизировать биллер (#3773)

+++ modules/uu/tarificator/AccountLogPeriodTarificator.php
@@ -7,8 +7,6 @@ use app\helpers\DateTimeZoneHelper;
 use app\modules\uu\classes\AccountLogFromToTariff;
 use app\modules\uu\models\AccountLogPeriod;
 use app\modules\uu\models\AccountTariff;
-use app\modules\uu\models\AccountTariffLog;
-use app\modules\uu\models\Period;
 use app\modules\uu\models\ServiceType;
 use RangeException;
 use Yii;
@@ -30,9 +28,12 @@ class AccountLogPeriodTarificator extends Tarificator
     public function tarificate($accountTariffId = null, $isWithTransaction = true)
     {
         $minLogDatetime = AccountTariff::getMinLogDatetime();
+        $now = date(DateTimeZoneHelper::DATE_FORMAT);
 
         // в целях оптимизации удалить старые данные
-        AccountLogPeriod::deleteAll(['<', 'date_to', $minLogDatetime->format(DateTimeZoneHelper::DATE_FORMAT)], [], 'id ASC');
+        if (!$accountTariffId) {
+            AccountLogPeriod::deleteAll(['<', 'date_to', $minLogDatetime->format(DateTimeZoneHelper::DATE_FORMAT)], [], 'id ASC');
+        }
 
         $accountTariffs = AccountTariff::find();
         $accountTariffId && $accountTariffs->andWhere(['id' => $accountTariffId]);
@@ -44,15 +45,28 @@ class AccountLogPeriodTarificator extends Tarificator
                 $this->out('. ');
             }
 
-            /** @var AccountTariffLog $accountTariffLog */
-            $accountTariffLogs = $accountTariff->accountTariffLogs;
-            $accountTariffLog = reset($accountTariffLogs);
+            /** @var AccountTariff $accountTariff */
+            $accountTariffLog = $accountTariff->getAccountTariffLogs()->one();
             if (!$accountTariffLog ||
                 (!$accountTariffLog->tariff_period_id && $accountTariffLog->actual_from_utc < $minLogDatetime->format(DateTimeZoneHelper::DATETIME_FORMAT))
             ) {
                 // услуга отключена давно - в целях оптимизации считать нет смысла
                 // @todo денормализовать услугу, чтобы хранить там эту дату и не лазить каждый раз в лог
                 continue;
+            }
+
+            if (!$this->isFullTarification) {
+                // сокращенная проверка
+                $accountLogPeriodDateTo = $accountTariff->getAccountLogPeriods()
+                    ->select(['date_to'])
+                    ->orderBy(['id' => SORT_DESC])
+                    ->scalar();
+                if ($accountLogPeriodDateTo && $accountLogPeriodDateTo > $now) {
+                    // уже есть списание абонентки на будущее. Дальше списывать нет необходимости
+                    // сравнение строго больше, чтобы не зависеть от таймзоны
+                    // @todo денормализовать услугу, чтобы хранить там эту дату и не лазить каждый раз в лог
+                    continue;
+                }
             }
 
             $isWithTransaction && $transaction = Yii::$app->db->beginTransaction();
